transcription-requirements-backup.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
huggingface-hub==0.31.2
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai-whisper==20240930
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
primePy==1.3
propcache==0.3.1
protobuf==6.31.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.41
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typer==0.15.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
